Rename accumulator to adder and use AddAssign instead of custom trait
We are completely reimagining the crate with this commit, centereing
folding types around existing traits instead of a custom one that
can't be implemented for builtin types anyway.

The wrapper type now defines what is used in FromIterator and Extend,
and it's always going to be an existing trait (for now).